refac (#78)

+++ tst/test/span.cpp
@@ -24,7 +24,7 @@ TEST(Span, init) {
 TEST(SpanSet, containsCorrectData) {
   mkn::kul::SpanSet<double> spanset{std::vector<size_t>{2, 2, 2}};
   mkn::kul::Span<double> raw = spanset.raw();
-  std::vector<size_t> vals{1, 5, 5, 3, 4, 5};
+  std::vector<double> vals{1, 5, 5, 3, 4, 5};
   for (size_t i = 0; i < 6; i++) raw[i] = vals[i];
   size_t i = 0;
   for (auto const& span : spanset)
